CCD-6018: Reduce data store production and AAT ES timeout (#39002)
* CCD-6018: Revert production and AAT ES timeout

Revert the Elastic Search timeout value for ccd-data-store-api back to 6 seconds in the production and AAT environments.  This was originally planned to be carried out under CCD-2275.

* CCD-6018: Revert production and AAT ES timeout

Reduced Elastic search timeout value to 12 seconds initially to allow for a period of monitoring with the new value.  If it looks OK then will reduce and monitor again.